improve number parsing
This correctly tokenizes numbers like `1e+` as a DIMENSION with the number
as `1` and the dimension as `e`, followed by a delim of `+`. Prior to
this change the lexer would consider `1e+` a number in it's entirety,
but this is incorrect per the spec:

> If the [next 2 or 3 input code points](https://drafts.csswg.org/css-syntax/#next-input-code-point) are U+0045 LATIN CAPITAL LETTER E (E) or U+0065 LATIN SMALL LETTER E (e), optionally followed by U+002D HYPHEN-MINUS (-) or U+002B PLUS SIGN (+), followed by a [digit](https://drafts.csswg.org/css-syntax/#digit), then:

Fixes #6

This change also introduces signCharacter which will be `+` or `-`, if
present.