fix number parsing for multiple dots
Prior to this change the lexer would parse numbers like `.2.7` as a
single number, due to greedily eating the first `.`. The result would be
a single number token, which is incorrect as consuming a number should
stop after seeing the second `.`

The fix is to not greedily consume the first `.`. This now means that
`.2.7` is correctly tokenized as two separate number tokens.